차트 못고침;;;
ㅠㅠ
