+++ src/main/resources/config/mybatis/mapper/oracle/chart.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" 
 "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.mybatis.mapper.oracle.chart">
+<mapper namespace="com.will.team4final.chart">
  	<select id="sumByMonth" parameterType="string" resultType="int">
 		select sum(price)as sum from payment
 		where substr(regdate, 4, 2) = #{month}
